+++ src/Canvas.tsx
@@ -20,7 +20,6 @@ type SharedDataGridProps<R, K extends keyof R> = Pick<DataGridProps<R, K>,
 | 'onSelectedCellChange'
 | 'onSelectedCellRangeChange'
 | 'onRowClick'
-| 'onRowDoubleClick'
 | 'onRowExpandToggle'
 | 'onSelectedRowsChange'
 > & Required<Pick<DataGridProps<R, K>,
@@ -205,7 +204,6 @@ function Canvas<R, K extends keyof R>({
           scrollLeft={nonStickyScrollLeft}
           isRowSelected={selectedRows?.has(row[rowKey]) ?? false}
           onRowClick={props.onRowClick}
-          onRowDoubleClick={props.onRowDoubleClick}
           onRowExpandToggle={props.onRowExpandToggle}
           enableCellRangeSelection={typeof props.onSelectedCellRangeChange === 'function'}
         />